Refactor sendScoreboardMessage to GameSession
